% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{track}
\alias{track}
\title{Tracks events object to database}
\usage{
track(event)
}
\arguments{
\item{event}{Object to save}
}
\description{
This function saves to the database behind the masters project api
}
